orchestrators/runpod/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
redis==5.2.0
docker==7.1.0
python-dotenv==1.0.0
